feat: add feature for group

+++ src/route/features/group/group.service.ts
@@ -47,11 +47,13 @@ export class GroupService {
     }
   }
 
-  async createGroup(body: CreateGroupDto): Promise<Group> {
+  async createGroup(body: CreateGroupDto, idAuthor: string): Promise<Group> {
     try {
       return this.groupRepository.save({
         id: uuid(),
         ...body,
+        author: idAuthor,
+        admins: [idAuthor],
       });
     } catch (error) {
       throw new BadRequestException('Error creating group');
@@ -60,6 +62,7 @@ export class GroupService {
 
   async updateGroup(
     id: string,
+    idUser: string,
     body: UpdateGroupDto,
   ): Promise<{ status: number; message: string }> {
     try {
@@ -68,6 +71,13 @@ export class GroupService {
       if (!group) {
         throw new NotFoundException('Group not found');
       }
+
+      const isAdmin = group.admins.includes(idUser);
+
+      if (!isAdmin) {
+        throw new BadRequestException('You are not admin of this group');
+      }
+
       group = { ...group, ...body };
       await this.groupRepository.save(group);
 
@@ -80,12 +90,20 @@ export class GroupService {
     }
   }
 
-  async deleteGroup(id: string): Promise<{ status: number; message: string }> {
+  async deleteGroup(
+    id: string,
+    idUser: string,
+  ): Promise<{ status: number; message: string }> {
     try {
       const group = await this.groupRepository.findOne({ where: { id } });
 
       if (!group) {
         throw new NotFoundException('Group not found');
+      }
+
+      const isAdmin = group.admins.includes(idUser);
+      if (!isAdmin) {
+        throw new BadRequestException('You are not admin of this group');
       }
       await this.groupRepository.remove(group);
 
@@ -95,6 +113,76 @@ export class GroupService {
       };
     } catch (error) {
       throw new BadRequestException(`Error deleting group`);
+    }
+  }
+
+  async joinGroup(
+    idGroup: string,
+    idUser: string,
+  ): Promise<{ status: number; message: string }> {
+    try {
+      const group = await this.groupRepository.findOne({
+        where: { id: idGroup },
+      });
+      if (!group) {
+        throw new NotFoundException('Group not found');
+      }
+      group.membersReq = [idUser, ...group.membersReq];
+      await this.groupRepository.save(group);
+      return {
+        status: HttpStatus.OK,
+        message: 'Join group successfully',
+      };
+    } catch (error) {
+      throw new BadRequestException('Error joining group');
+    }
+  }
+
+  async acceptMemberReqJoinGroup(
+    idGroup: string,
+    idUser: string,
+    idMemberReq: string,
+  ): Promise<{ status: number; message: string }> {
+    try {
+      const group = await this.groupRepository.findOne({
+        where: { id: idGroup },
+      });
+      if (!group) {
+        throw new NotFoundException('Group not found');
+      }
+      group.members = [idUser, ...group.members];
+      group.membersReq = group.membersReq.filter(
+        (i: string) => i !== idMemberReq,
+      );
+      await this.groupRepository.save(group);
+      return {
+        status: HttpStatus.OK,
+        message: 'Accept member req join group successfully',
+      };
+    } catch (error) {
+      throw new BadRequestException('Error accept member req join group');
+    }
+  }
+
+  async leaveGroup(
+    idGroup: string,
+    idUser: string,
+  ): Promise<{ status: number; message: string }> {
+    try {
+      const group = await this.groupRepository.findOne({
+        where: { id: idGroup },
+      });
+      if (!group) {
+        throw new NotFoundException('Group not found');
+      }
+      group.members = group.members.filter((i: string) => i !== idUser);
+      await this.groupRepository.save(group);
+      return {
+        status: HttpStatus.OK,
+        message: 'Leave group successfully',
+      };
+    } catch (error) {
+      throw new BadRequestException('Error leaving group');
     }
   }
 }
